refactored ExportThread
